refactor: extract finalize + rematch and complete spine split
Move finalizeReview into wvImportFinalizeController; move
rematchWorldView, runRematch helper, and getRematchStatus into
wvImportRematchController. The worldViewImportController monolith is
now a pure barrel re-exporting from 8 focused domain files plus the
wvImportSharedState module.

Final result of the 8-PR spine domain split (ADR-0009):
- wvImportLifecycleController       (PR-10)
- wvImportMatchController           (PR-11)
- wvImportCoverageController        (PR-12)
- wvImportAIController              (PR-13)
- wvImportTreeOpsController         (PR-14)
- wvImportHierarchyController       (PR-15)
- wvImportFlattenController         (PR-16)
- wvImportFinalizeController        (PR-17, this commit)
- wvImportRematchController         (PR-17, this commit)
- wvImportSharedState (state)       (introduced PR-14)

Pure refactor — no behavior change, byte-identical function bodies.
adminRoutes.ts is unchanged across all 8 PRs.
